btrfs: introduce sealed subvolume
A 'sealed' subvolume means that once the subvolume is set read-only, it
can never change back (under normal circumstances, through ioctls).

This should prevent any accidental change of a subvolume and is intended
to be used for send/receive.

A sealed subvolume:
* appears as any other read-only subvolume
* can be created directly via snapshot ioctl
* can be marked as sealed from any ro/rw subvolume
* can be snapshotted again without restrictions
* cannot be made RW again

For this purpose a new flag bit is used along with SUBVOL_RDONLY, does
not change on-disk format. Any older kernel will not honor the SEALED
flag. Off-line modification of the flag is possible and cannot be
prevented.

Signed-off-by: David Sterba <dsterba@suse.cz>