btrfs: separate mount options that could be set per-subvolume
Introduce the data structure that holds the set mount options and use it
for filesystem mount options for now. Use it to store per-option data.

We have to additionally track if a given option was set or not, it's not
possible to find that out from current mount_opt. This will be used
later to inherit per-filesystem option if the per-subvolume weren't set.

Prepare for compress_type and max_inline.

The mount options are still applied to the whole filesystem, no visible
change.

Signed-off-by: David Sterba <dsterba@suse.cz>