btrfs: enum the set of per-subvolume mount options
The set of currently supported per-subvolume mount options is:

* nodatasum, nodatacow
* compress, compress-force
* max_inline

Note: remounting the subvolume with different flags does not work. The
VFS inteface passes only superblock, we need the subvolume root. This
would require changes in VFS so remount will work only fs-wide.

Mounting a given subvolume again will inherit the options from the
previous mount. IOW, all mounts of a given subvolume share the same
mount options. A new mount of a given subvolume will change the options
so this is a funny way how to simulate the remount.

Options that could be supported, but there are some problems left:

* inode_cache - pending options would require to track all roots that
  got changed by remount - but we know that remount does not work anyway

* autodefrag - again remount, but here we want to flush all scheduled
  defrags before remount and we don't get the root

Both could be added later. Autodefrag is easier, we'd just go through
all roots and flush their defrag work. The utility of inode_cache
option is questinable on itself.

Signed-off-by: David Sterba <dsterba@suse.cz>