btrfs: initialize temporary mount options before parsing
The tree_root serves as temporary storage of the mount options. All
changes have to be transfered to the respective mount option owner as
soon as it's known.

The way the mount works, we don't have the pointer to the mounted
subvolume (subvol= or subvolid=), so we have to store it somewhere. The
API to set/clear the mount options stays, we just have to keep track of
the changes via tree_root.

Signed-off-by: David Sterba <dsterba@suse.cz>