btrfs: remove unlikely from ok-ish static predictions
The compiler's internal heuristics predict that values are unlikely to
be zero when checked, the annotations are redundant and we generally
don't want to encourage them in such cases. The runtime effects are
non-measurable.

The resulting assembly code is not the same as before this patch, but
the it does not seem to be significantly changing the code, some blocks
got reordered or shifted.

Signed-off-by: David Sterba <dsterba@suse.com>