btrfs: switch auth key keyring to 'user'
The 'logon' keyring is hardcoded and keys in that keyring have disabled
reading the bytyes from userspace, for security reasons. This makes it
hard to test the feature as mkfs cannot access the key and it must be
specified in another way.

Future plans: specify the keyring as part of the key name, to allow more
than just user or logon keyrings or ways how to pass the key.

Signed-off-by: David Sterba <dsterba@suse.com>