btrfs: update RAID1 and RAID10 calculations in statfs
The RAID1* profiles don't need to be enumerated, we can use the mask.

For RAID10 we need to take into account the degenerate mode, the
redundancy level is also 'ncopies' and that like RAID0, it spans all the
devices. There's clearly a maximum of nr_devices that the allocator can
take into account. Using 4 here means the non-degenerate minimum number
of devices (later adjusted to the nr_devices if smaller), so the
redundancy is not important here, we always want to simulate using the
full stripe, so it's nr_devices for RAID10.

Signed-off-by: David Sterba <dsterba@suse.com>