btrfs: allow defrag to be interruptible
During defrag, at btrfs_defrag_file(), we have this loop that iterates
over a file range in steps no larger than 256K subranges. If the range
is too long, there's no way to interrupt it. So make the loop check in
each iteration if there's a fatal signal pending, and if there is, break
and return -EINTR to userspace.

This is motivated by a recent bug on 5.16 where defragging a 1 byte file
resulted in iterating from file range 0 to (u64)-1, as hitting the bug
triggered a too long loop, basically requiring one to reboot the machine.

Signed-off-by: Filipe Manana <fdmanana@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>