btrfs: store details about first setget bounds check failure
The way the setget helpers are used makes it hard to return an error
code at the call site, as this would require to clutter a lot of code
with potential failures that are expected to be rare.

Instead, do a delayed reporting, tracked by a filesystem-wide bit that
synchronizes potential races in the reporting function that records only
the first event. To give a bit more insight into the scale, count the
total number of events.

Signed-off-by: David Sterba <dsterba@suse.com>