btrfs: make callers explicitly request bio submission
At present, submit_extent_page() takes a page and queues it for
submission to the underlying device on the provided bio_ctrl object. If
the page is not able to be added on the bio_ctrl's current bio, the
current bio is submitted and a new bio is allocated, after which the
page must be able to be added to the new bio. The function also has a special-purpose parameter, force_bio_submit, which is used only
by compressed IO, in order to force the existing bio in the bio_ctrl
object to be submitted before adding the new page.

An alternate approach, beyond having the special-purpose parameter, is
to explicitly provide a function to submit the currently pending bio in
the bio_ctrl object, if there is one. This removes the need for the
parameter force_bio_submit, and makes btrfs_readpage somewhat more
elegant also. TODO addc more dexription here, split up change better