btrfs: explicitly keep track of file extent item size.
Previously, file extents were always of constant size. However, for
fscrypt, they will need to store an IV, and this IV varies in length
according to the encryption algorithm, making the file extent items also
of variable length. Therefore, this begins passing around the actual
file extent item size instead of assuming it is merely the size of the
item in memory.

Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>