fscrypt: add fscrypt_have_same_policy() to check inode's compatibility
btrfs will have the possibility of encrypted and unencrypted files in
the same directory, and it's important to not allow these two files to
become linked together. Therefore, add a function which allows checking
the encryption policies of two inodes to ensure they are compatible.

TODO: This might not be necessary if we disallow other policies other
than the IV_FROM_FS policy for btrfs; it's currently unused.

Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>