fscrypt: add filename-like encryption of sub-page chunks
Many filesystems have the notion of an inline extent, where a
sufficiently small file can have its data stored directly in its extent
instead of stored in its own page and ultimately disk block. However,
currently encryption precludes data inlining, and the data is forced to
occupy a whole block on disk.

Others implement fs-verity to ensure files are secure; however, verity
stores hashes of the data, and needs to be able to lookup data blocks by
hash. Obviously, if a file is encrypted, the hashes should reflect
unencrypted data, but should be themselves encrypted. However, the
hashes rarely line up with a full page and need somewhat random
addressibility.

This change adds facilities for a filesystem to request the
encryption/decryption of a sub-page chunk of data; this will allow
inline data extent encryption as well as encryption of verity hashes.

Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>