btrfs: add fscrypt_context items
In order to store encryption information for directories, symlinks,
etc., fscrypt with per-extent keys stores a context item with each
encrypted non-regular inode. fscrypt provides an
arbitrary blob for the filesystem to store, and it does not clearly fit
into an existing structure, so this goes in a new item type.

Signed-off-by: Omar Sandoval <osandov@osandov.com>
Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>