btrfs: add iv generation function for fscrypt
As btrfs cannot use the standard inode or logical block based encryption
for data block encryption, it must provide a IV generation function and
users must use the IV_FROM_FS policy. For filenames, we can just use the
nonce that fscrypt stores per-inode, since these encrypted datum are not
shared between inodes; later on, we will store an IV per file extent,
and return it in this function for data encryption.

Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>