fscrypt: expose fscrypt_nokey_name
btrfs needs to deal with matching names using its own methods, as it
uses methods on a struct extent_buffer for comparison of names instead
of working against a raw byte array. Thus, it needs to deal in
nokey_names just as fscrypt does, and cannot do that unless the
structure is exposed.

Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>