fscrypt: Add new encryption policy for btrfs.
Certain filesystems want to generate an IV for a particular block in
some way other than by inode number plus block number. In particular,
btrfs can change the block number for a particular piece of data, and
can have multiple inodes using the same block, so to use a IV generated
from the inode or the block number would interfere with normal
filesystem operation.

In order to prevent needing to plumb arbitrary methods of IV generation
into fscrypt, this change adds a new policy, IV_FROM_FS. If this policy
is used, the filesystem is required to provide an operation function
pointer, get_fs_derived_iv, which takes the inode and the offset of the
block within the inode and fills in the IV buffer itself. Such a
filesystem is expected to appropriately generate and store a persistent
random IV for each block of data.

Signed-off-by: Sweet Tea Dorminy <sweettea-kernel@dorminy.me>